+++ .gitignore
@@ -3,10 +3,9 @@
 # dependencies
 /node_modules
 **/node_modules
-**/**/node_modules
 package-lock.json
 **/package-lock.json
-**/**/package-lock.json
+
 # testing
 /coverage
