+++ package.json
@@ -10,7 +10,7 @@
     "link:angular": "bash ./linkAngularMapbox.sh",
     "start:angular": "cd angular-demo && yarn && cd .. && yarn link:angular && cd angular-demo && yarn start",
     "link:react": "bash ./linkReactMapbox.sh",
-    "start:react": "cd react-demo && yarn && cd .. && yarn link:react && cd react-demo && yarn start" 
+    "start:react": "cd react-demo && yarn && cd .. && yarn link:react && cd react-demo && yarn start"
   },
   "repository": {
     "type": "git",
@@ -25,5 +25,8 @@
   "bugs": {
     "url": "https://github.com/pxblue/mapbox/issues"
   },
-  "homepage": "https://github.com/pxblue/mapbox#readme"
+  "homepage": "https://github.com/pxblue/mapbox#readme",
+  "dependencies": {
+    "react-app-polyfill": "^1.0.2"
+  }
 }
