+++ react-demo/README.md
@@ -1,12 +1,12 @@
-# PX Blue Mapbox
-This is a sample application demonstrating the appearance of the PX Blue themes.
+# PX Blue Mapbox Demo
+This package contains various Mapbox themes for use with PX Blue. 
 
 | Framework           | Status       | Live Examples  |
 | ---------------- |--------------|------------------|
-| Angular | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/angular.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/angular) | [View on Stackblitz](https://stackblitz.com/edit/pxblue-mapbox-angular)
-| React | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/react.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/react) | [View on Stackblitz](https://stackblitz.com/edit/pxblue-mapbox-react)
-| Ionic | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/ionic.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/ionic) | [View on Stackblitz](https://stackblitz.com/edit/pxblue-mapbox-ionic)
-| React Native | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/reactnative.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/reactnative) | [View on Snack](https://snack.expo.io/@px-blue/mapbox-reactnative)
+| Angular | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/angular.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/angular) | [View on Stackblitz](https://stackblitz.com/github/pxblue/mapbox/tree/angular)
+| React | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/react.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/react) | [View on Code Sandbox](https://codesandbox.io/s/github/pxblue/mapbox/tree/react)
+| Ionic | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/ionic.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/ionic) | [View on Stackblitz](https://stackblitz.com/github/pxblue/mapbox/tree/ionic)
+| React Native | [![CircleCI](https://circleci.com/gh/pxblue/mapbox/tree/reactnative.svg?style=shield)](https://circleci.com/gh/pxblue/mapbox/tree/reactnative) | [View on Snack](https://snack.expo.io/@git/github.com/pxblue/mapbox@reactnative)
 
 ## Running the Example
 ### Angular
@@ -33,6 +33,7 @@ yarn && ionic serve
 ```
 To run on a device, add the platform then build and run the project:
 ```
+ionic cordova resources
 ionic cordova platform add [android|ios]
 ionic cordova build [android|ios]
 ionic cordova run [android|ios]
